getopt 설정 완료
